refactor: move run polling to client package
